requirements.txt
streamlit
numpy
pandas
scikit-learn
xgboost
yfinance
matplotlib
seaborn
pypfopt
json5
pillow